backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0
PyMySQL==1.1.0
bcrypt==4.1.2
python-dotenv==1.0.0
Werkzeug==3.0.1
Pillow==10.2.0
gunicorn==21.2.0
cloudinary==1.40.0
